refactor(pi_cli): co-locate pi_* modules under yoyopod_cli/pi/
Moves pi_network.py, pi_voip.py, pi_power.py, and the pi_validate/
subpackage built earlier in this PR into a single yoyopod_cli/pi/
package whose layout mirrors the CLI shape:

    yoyopod_cli/pi/
    ├── __init__.py    # typer.Typer assembly for the `pi` command group
    ├── network.py
    ├── voip.py
    ├── power.py
    └── validate/      # the subpackage from earlier in this PR

The typer assembly previously inline in main.py (lines 128-140 imports
plus 4 add_typer calls) collapses to one import + one add_typer:

    from yoyopod_cli.pi import app as pi_app
    app.add_typer(pi_app, name="pi")

main.py loses 13 lines of CLI plumbing.

Pure mechanical refactor — same CLI surface (`yoyopod pi <subcommand>`
unchanged), same flags, same exit codes. Internal imports inside
pi/validate/ updated via sed (yoyopod_cli.pi_validate ->
yoyopod_cli.pi.validate). 7 test files updated for the new submodule
paths via sed sweep.
