+++ yoyopod_cli/pi/validate/_navigation_soak/idle.py
