+++ yoyopod_cli/pi/validate/lvgl.py
@@ -1,4 +1,4 @@
-"""Stability validation subcommand."""
+"""Lvgl validation subcommand."""
 
 from __future__ import annotations
 
@@ -8,16 +8,19 @@ import typer
 
 from yoyopod_cli.common import configure_logging
 from yoyopod_cli.defaults import DEFAULT_TEST_MUSIC_TARGET_DIR
-from yoyopod_cli.pi_validate._navigation_soak import (
+from yoyopod_cli.pi.validate._navigation_soak import (
     NavigationSoakError,
     run_navigation_idle_soak,
 )
 
 
-def stability(
+def lvgl(
     config_dir: Annotated[
         str, typer.Option("--config-dir", help="Configuration directory to use.")
     ] = "config",
+    simulate: Annotated[
+        bool, typer.Option("--simulate", help="Run against simulation instead of hardware.")
+    ] = False,
     cycles: Annotated[
         int, typer.Option("--cycles", help="How many full transition cycles to run.")
     ] = 2,
@@ -33,14 +36,14 @@ def stability(
         bool,
         typer.Option(
             "--with-music",
-            help="Also exercise playlist loading and now-playing actions during the soak.",
+            help="Exercise playlist loading and now-playing actions during the soak.",
         ),
     ] = False,
     provision_test_music: Annotated[
         bool,
         typer.Option(
             "--provision-test-music/--no-provision-test-music",
-            help="Seed deterministic validation music before playback soak steps.",
+            help="Seed the deterministic validation music library before playback soak steps.",
         ),
     ] = True,
     test_music_dir: Annotated[
@@ -51,17 +54,19 @@ def stability(
         ),
     ] = "",
     skip_sleep: Annotated[
-        bool, typer.Option("--skip-sleep", help="Skip the sleep and wake exercise.")
+        bool, typer.Option("--skip-sleep", help="Skip the sleep/wake exercise.")
     ] = False,
     verbose: Annotated[bool, typer.Option("--verbose", help="Enable DEBUG logging.")] = False,
 ) -> None:
-    """Run a repeated navigation and idle stability pass on the target checkout."""
+    """Run a deterministic LVGL navigation and idle soak pass against YoYoPod."""
+    from loguru import logger
+
     configure_logging(verbose)
     resolved_music_dir = test_music_dir or DEFAULT_TEST_MUSIC_TARGET_DIR
     try:
         report = run_navigation_idle_soak(
             config_dir=config_dir,
-            simulate=False,
+            simulate=simulate,
             cycles=cycles,
             hold_seconds=hold_seconds,
             idle_seconds=idle_seconds,
@@ -71,11 +76,6 @@ def stability(
             test_music_dir=resolved_music_dir,
         )
     except NavigationSoakError as exc:
-        from loguru import logger
-
-        logger.error(f"Stability soak failed: {exc}")
+        logger.error(f"LVGL soak failed: {exc}")
         raise typer.Exit(code=1)
-
-    from loguru import logger
-
-    logger.info(f"Stability soak passed: {report.summary()}")
+    logger.info(f"LVGL soak passed: {report.summary()}")
